Fit endpoint returns long-running Operation
POST requests to the Fit endpoint return an Operation. This is
appropriate because (non-cached) Fits will be long-running. Client can
make requests to check on the status of an Operation. When the operation
is done, the client knows that it's safe to retrieve the

This commit also removes unneeded `Error` schema.

Better conforms to Google API Design Guide. The `Error` schema was
incorrectly adopted. It's part of a Google-specific
backward-compatibility scheme. We do not need backward-compatibility.
See https://cloud.google.com/apis/design/errors for details about the
schemas.