fix: move SPSCQueue into anon stan model template
This makes the code slightly more opaque but there's a very good reason
for doing this. It is not possible to share `std::string`s across Python
extension models compiled against different C++ `stdlib`s.

One alternative is to use char * instead of std:string but this is
unattractive because memory management is required.