fix: Disable gcc testing, due to Stan math 2.20 incompatibility
Building a Python extension module using Stan math 2.20 with STAN_THREADS on results
in an immediate segfault when using gcc (but not clang) on Linux. The problem may
involve a static initialization order problem.

Since STAN_THREADS is required, the short-term fix is to drop support
for gcc. The problem occurs with all versions of gcc, including gcc 9.