feat: Encode Stan writer output using protobuf in C++
Previously Python was used. Doing everything directly in C++ is much
faster and we eliminate the dependency on the Python protobuf library.

This version requires that the user install protobuf system-wide. Once
this solution is verified as working, we can remove this limitation by
copying the shared library into place.