fix: Run stan::services functions in background using ProcessPool
Use a ProcessPoolExecutor instead of a ThreadPoolExecutor when
calling long-running stan::services functions.

Virtually all the changes in this commit follow from this change.
Switching to use a Unix domain socket to pass information from
subprocesses to the main process is required because the previous
method involved sharing memory between threads. Since we no longer
use threads, we cannot share memory in this way.

Remove the `PRAGMA journal_mode=WAL` setting for sqlite3 to avoid issues
associated with accessing the sqlite3 database from different processes.

Since we cannot use Windows currently due to the use of Unix domain sockets,
remove the custom `TemporaryDirectory` in favor of the standard library version.