fix: Use stream instead of datagram over socket
Sending datagram messages (à la UDP) is extremely convenient. The code
was readable. Unfortunately, macOS has a very small send buffer. The
Stan logger and writer callbacks (which send us draws) send a lot of
messages and exceed the capacity of the buffer on macOS.

So we use the stream protocol instead. Stan will have to wait until we
are ready to send us draws.