test: "Resolve" flaky tests under clang
When `clang` is used as the compiler, several tests are flaky. That is,
they fail when run with the entire test suite but pass when run by
themselves.

These flaky tests disappear if every test is run in a separate process.
There is a pytest plugin which does this. We use it when `clang` is used
as the compiler.