feat: Build services module in cache destination
Build the model-specific stan::services extension module in the cache
directory. This eliminates the need to save the module in a database and
copy it into a temporary directory every time it is used.

This commit also renames some functions to make clear the distinction
between the extension module (which wraps stan::services functions)
and the C++ Stan model (generated by stanc).