feat: Remove Cython build dependency
Remove inconvenient build-time dependency on Cython.
Cython is used by httpstan to wrap the model-specific stan::services
functions. Cython was also used to build a module "empty" whose
only purpose was to force the Python wheel building tools to
recognize the httpstan wheel as platform-specific (which it is due
to the presence of shared libraries in `httpstan/lib`).

Closes #399