fix: Performance improvement in `select` loop
Make minor adjustment which seem to help with performance.
The `select` loop in services_stub is reading messages from Stan
which can come in extraordinarily quickly or very slowly depending
on how complicated the model is. If they are coming in quickly we need
to read them as quickly as possible.

Setting the timeout to 0.01 in the call to `select` means
spending a bit more time waiting for the next message rather
than `sleep`ing via the asyncio loop. Odds are that what we want is
the next message from Stan, so making the timeout 0.01 rather than 0
means we will spend more time waiting for what we likely want.

This, at least, is how I explain the fact that this seems to help a bit
empirically.