refactor: Align Cython code better with CmdStan
Make minor changes to better align Cython code with Cmdstan's
command.hpp. If and when people are interested in understanding how this
wrapping code works it will be helpful to have all code which
does the same thing look approximately the same. For example,
CmdStan references the model with an instance of model_base, so now we
do as well. No changes here change behavior.