feat: Add list_models endpoint
Allows users to list models in cache. REST APIs
typically offer such a list.

Such a list is required if a user wants to delete
all models from the cache. This is because, without
this function, there is no way for the user to know
what models are in the cache. Models must be referred
to by name in order to delete them.

Closes #409