BREAKING CHANGE: Use bytes not strings in protobuf messages
Use `bytes` not `string`s in Protocol Buffers messages. According to the
protocol buffers specification, `string`s may only be used when one can
guarantee that the underlying bytes are valid UTF-8.  (We use C++
`std::string`s and perform no checking that anything is valid UTF-8.)
According to the spec, we must use Protocol Buffer `bytes`.

This may address a mysterious bug on macOS that seems to involve
a failure in the encoding (or decoding) of a protocol buffer message.

This is a breaking change since we are changing the protocol buffer
schema.