feat: Compress draws and messages using LZ4
Compress draws and messages from Stan using LZ4.  Previously gzip was
used. (LZ4 is much faster than gzip.) This commit also removes
compression/decompression steps from the `load_fit` and `dump_fit`
functions. Cache functions should be as simple as possible. In
particular, they should not need to know the format of data being
stored.  That said, an indication of format remains in the filename
`jsonlines.lz4`.  This is unfortunate and should be addressed in a
future commit.

Closes #485