feat: Add log_prob_grad method and endpoint
Given a model name, associated data and unconstrained parameters,
the log_prob_grad endpoint will calculate and return the gradient
of the unconstrained parameters.

This feature is accompanied by a test: the log_prob_grad endpoint is
validated on a simple Gaussian example, the test compares the output
of the endpoint against an analytical calculation of the gradient.