fix: Do not save messages from failed fits
Do not save any messages from failed sampling runs.
For example, if sampling starts but cannot continue because attempts
to find appropriate initial values are rejected, there will be some
(logger) messages saved to disk. The presence of these messages
confuses the cache machinery, which assumes the presence of messages
on disk indicates a successful fit. Deleting them avoids the problem.

There remains the problem of retrieving the logger messages associated
with the failed fit. It's possible that this is something someone might
want to do. Future work might address this. Such a user is also free to
use CmdStan.

Closes #452