Move "fit" ProcessPool out of module top-level
Creating the ProcessPoolExecutor at the top level of the services_stub.py
is a problem because the pool can't be shut down properly. This can break
code using the httpstan library because the processes are never shut down
and the calling process might hang on exit.

I think this change solves the issue that creating the pool at the module
top-level was meant to solve, but does so in a way that allows the process
pool to complete its lifecycle, thus playing well with others.